Backend/requirements.txt
Django==2.2.18
django-filter==2.0.0
djangorestframework==3.11.2